# ENH: single-pass CSR sparse matrix binops