@rgommers 2023-12-28
Thanks @alugowski!

First benchmark run (same as on gh-19766) seems to show that things are indeed significantly faster, except for the DIA format. 
```
| Change   | Before [bee9b9ae] <main>   | After [f427cf01] <fastbinop>   |   Ratio | Benchmark (Parameter)                                      |
|----------|----------------------------|--------------------------------|---------|------------------------------------------------------------|
| +        | 47.1±0.7ms                 | 54.9±5ms                       |    1.17 | sparse.Arithmetic.time_arithmetic('dia', 'BB', '__sub__')  |
| +        | 67.1±1ms                   | 77.9±7ms                       |    1.16 | sparse.Arithmetic.time_arithmetic('dia', 'BB', 'multiply') |
| +        | 67.8±0.2ms                 | 77.3±7ms                       |    1.14 | sparse.Arithmetic.time_arithmetic('dia', 'BB', '__add__')  |
| -        | 72.0±7ms                   | 64.6±1ms                       |    0.9  | sparse.Arithmetic.time_arithmetic('csr', 'AB', '__sub__')  |
| -        | 162±7ms                    | 143±3ms                        |    0.88 | sparse.Arithmetic.time_arithmetic('coo', 'AB', 'multiply') |
| -        | 174±20ms                   | 145±1ms                        |    0.84 | sparse.Arithmetic.time_arithmetic('coo', 'AB', '__sub__')  |
| -        | 25.7±2ms                   | 21.1±0.09ms                    |    0.82 | sparse.Arithmetic.time_arithmetic('csr', 'AA', 'multiply') |
| -        | 23.3±2ms                   | 18.9±0.4ms                     |    0.81 | sparse.Arithmetic.time_arithmetic('csc', 'AA', '__sub__')  |
| -        | 76.3±8ms                   | 61.0±2ms                       |    0.8  | sparse.Arithmetic.time_arithmetic('csr', 'AB', 'multiply') |
| -        | 26.8±3ms                   | 18.3±0.2ms                     |    0.68 | sparse.Arithmetic.time_arithmetic('csr', 'AA', '__add__')  |
```

These benchmarks need a bit of care, because the result is now going to depend strongly on whether the result is in canonical form or not. Here, it turns out that `A` is canonical while `B` is not, due to it being constructed as `A ** 2`.

If we insert something like this in the benchmark setup code to ensure both matrices are in canonical form:
```python
        if hasattr(x, 'sum_duplicates'):
            x.sum_duplicates()
        if hasattr(self.y, 'sum_duplicates'):
            self.y.sum_duplicates()
```
we get:
```
| Change   | Before [bee9b9ae] <main>   | After [f427cf01] <fastbinop>   |   Ratio | Benchmark (Parameter)                                      |
|----------|----------------------------|--------------------------------|---------|------------------------------------------------------------|
| +        | 28.4±0.3ms                 | 40.1±0.7ms                     |    1.41 | sparse.Arithmetic.time_arithmetic('csr', 'BB', '__sub__')  |
| +        | 28.5±0.4ms                 | 39.1±0.3ms                     |    1.37 | sparse.Arithmetic.time_arithmetic('coo', 'BB', '__sub__')  |
| +        | 28.3±0.2ms                 | 38.9±0.3ms                     |    1.37 | sparse.Arithmetic.time_arithmetic('dia', 'BB', '__sub__')  |
| +        | 30.2±1ms                   | 39.5±0.4ms                     |    1.31 | sparse.Arithmetic.time_arithmetic('csc', 'BB', '__sub__')  |
| +        | 50.0±2ms                   | 64.6±3ms                       |    1.29 | sparse.Arithmetic.time_arithmetic('csr', 'AB', '__sub__')  |
| +        | 46.8±2ms                   | 60.0±2ms                       |    1.28 | sparse.Arithmetic.time_arithmetic('csr', 'AB', 'multiply') |
| +        | 51.4±0.1ms                 | 64.5±1ms                       |    1.26 | sparse.Arithmetic.time_arithmetic('csc', 'BA', '__sub__')  |
| +        | 51.3±3ms                   | 64.8±1ms                       |    1.26 | sparse.Arithmetic.time_arithmetic('csr', 'AB', '__add__')  |
| +        | 52.7±2ms                   | 66.0±1ms                       |    1.25 | sparse.Arithmetic.time_arithmetic('csc', 'AB', '__sub__')  |
| +        | 50.8±2ms                   | 63.7±2ms                       |    1.25 | sparse.Arithmetic.time_arithmetic('csr', 'BA', '__sub__')  |
| +        | 54.7±0.8ms                 | 67.7±0.9ms                     |    1.24 | sparse.Arithmetic.time_arithmetic('csc', 'AB', 'multiply') |
| +        | 62.5±1ms                   | 76.9±1ms                       |    1.23 | sparse.Arithmetic.time_arithmetic('coo', 'AB', '__add__')  |
| +        | 54.3±0.5ms                 | 66.8±0.9ms                     |    1.23 | sparse.Arithmetic.time_arithmetic('csc', 'BA', 'multiply') |
| +        | 48.2±2ms                   | 58.7±3ms                       |    1.22 | sparse.Arithmetic.time_arithmetic('csr', 'BA', 'multiply') |
| +        | 60.3±0.9ms                 | 73.0±2ms                       |    1.21 | sparse.Arithmetic.time_arithmetic('coo', 'AB', 'multiply') |
| +        | 52.3±3ms                   | 63.4±2ms                       |    1.21 | sparse.Arithmetic.time_arithmetic('csr', 'BA', '__add__')  |
| +        | 62.3±2ms                   | 74.6±2ms                       |    1.2  | sparse.Arithmetic.time_arithmetic('coo', 'AB', '__sub__')  |
| +        | 61.9±1ms                   | 73.6±2ms                       |    1.19 | sparse.Arithmetic.time_arithmetic('coo', 'BA', '__sub__')  |
| +        | 54.2±0.8ms                 | 64.5±1ms                       |    1.19 | sparse.Arithmetic.time_arithmetic('csc', 'AB', '__add__')  |
| +        | 54.0±0.6ms                 | 64.0±0.4ms                     |    1.19 | sparse.Arithmetic.time_arithmetic('csc', 'BA', '__add__')  |
| +        | 65.3±4ms                   | 75.8±2ms                       |    1.16 | sparse.Arithmetic.time_arithmetic('coo', 'BA', 'multiply') |
| +        | 210±6ms                    | 233±7ms                        |    1.11 | sparse.Arithmetic.time_arithmetic('dia', 'BA', '__sub__')  |
| +        | 67.4±2ms                   | 74.2±2ms                       |    1.1  | sparse.Arithmetic.time_arithmetic('coo', 'BA', '__add__')  |
| +        | 209±5ms                    | 222±8ms                        |    1.06 | sparse.Arithmetic.time_arithmetic('dia', 'AB', '__add__')  |
| +        | 213±6ms                    | 225±7ms                        |    1.06 | sparse.Arithmetic.time_arithmetic('dia', 'BA', 'multiply') |
| -        | 59.3±2ms                   | 53.9±0.6ms                     |    0.91 | sparse.Arithmetic.time_arithmetic('coo', 'BB', '__add__')  |
| -        | 60.4±0.9ms                 | 54.6±0.6ms                     |    0.9  | sparse.Arithmetic.time_arithmetic('csc', 'BB', '__add__')  |
| -        | 59.3±0.5ms                 | 53.4±0.1ms                     |    0.9  | sparse.Arithmetic.time_arithmetic('dia', 'BB', '__add__')  |
| -        | 53.1±2ms                   | 47.1±2ms                       |    0.89 | sparse.Arithmetic.time_arithmetic('coo', 'AA', '__add__')  |
| -        | 26.8±0.6ms                 | 23.8±0.2ms                     |    0.89 | sparse.Arithmetic.time_arithmetic('csc', 'AA', 'multiply') |
| -        | 26.9±0.3ms                 | 20.9±0.2ms                     |    0.78 | sparse.Arithmetic.time_arithmetic('csc', 'AA', '__add__')  |
| -        | 23.5±0.3ms                 | 18.3±0.2ms                     |    0.78 | sparse.Arithmetic.time_arithmetic('csr', 'AA', '__add__')  |
```

`B` should be in canonical form there, but the results hint at the new single-pass check for that failing for `__sub__` and `multiply`. At least, otherwise I can't explain why those results regress here. I'd also expect the `__sub__` and `__add__` results to be symmetric, so something isn't quite right yet it looks like.

There's also a all of `-Wmaybe-uninitialized` compiler warnings - it'd be good to address those, and perhaps that catches something.

Running the `bench_sparse_add.py` script from the PR description, the results improve but not as much as in the PR description, more like in the 0% - 10% range for me:

<details>

```
random-50,000:  0.00027637499442789704
poisson-49,600:         0.00023411099391523749
random-500,000:         0.001635348002309911
poisson-498,016:        0.0017717820010147989
random-5,000,000:       0.023091215000022203
poisson-4,996,000:      0.02406818899908103
random-50,000,000:      0.22385444400424603
poisson-49,978,572:     0.23443531700468156

# main:
random-50,000:  0.00038524300907738507
poisson-49,600:         0.00021164999634493142
random-500,000:         0.00149274300201796
poisson-498,016:        0.0016441019979538396
random-5,000,000:       0.022221983992494643
poisson-4,996,000:      0.025533260995871387
random-50,000,000:      0.21116333699319512
poisson-49,978,572:     0.2499628399964422

# main run 2:
random-50,000:  0.00038719200529158115
poisson-49,600:         0.00022731799981556833
random-500,000:         0.001550414992379956
poisson-498,016:        0.0016773520037531853
random-5,000,000:       0.022551914007635787
poisson-4,996,000:      0.02617731600184925
random-50,000,000:      0.21719301299890503
poisson-49,978,572:     0.2543543719948502

# this PR, run 2:
random-50,000:  0.0002743800141615793
poisson-49,600:         0.00022587399871554226
random-500,000:         0.0016353649989468977
poisson-498,016:        0.0016516460018465295
random-5,000,000:       0.02321807699627243
poisson-4,996,000:      0.023670254988246597
random-50,000,000:      0.2263502830028301
poisson-49,978,572:     0.23783767299028113
```

</details>

I'm not yet sure what to make of all this. Maybe you can run the same asv benchmark @alugowski and see how that looks for you?