@perimosocordiae 2024-05-15
ASV benchmarks without the changes to ensure canonical format ahead of time:

| Change   | Before [8ce393f1] <main>   | After [d9a04dc7] <fastbinop>   |   Ratio | Benchmark (Parameter)                                      |
|----------|----------------------------|--------------------------------|---------|------------------------------------------------------------|
| +        | 3.48±0ms                   | 4.04±0ms                       |    1.16 | sparse.Arithmetic.time_arithmetic('dia', 'BB', '__add__')  |
| +        | 3.58±0ms                   | 4.07±0ms                       |    1.14 | sparse.Arithmetic.time_arithmetic('dia', 'BB', 'multiply') |
| +        | 4.97±0ms                   | 5.36±0ms                       |    1.08 | sparse.Arithmetic.time_arithmetic('csc', 'BA', '__add__')  |
| +        | 13.6±0ms                   | 14.3±0ms                       |    1.05 | sparse.Arithmetic.time_arithmetic('csc', 'BA', '__mul__')  |
| -        | 9.72±0ms                   | 9.19±0ms                       |    0.95 | sparse.Arithmetic.time_arithmetic('coo', 'AB', '__add__')  |
| -        | 11.4±0ms                   | 10.7±0ms                       |    0.94 | sparse.Arithmetic.time_arithmetic('coo', 'AA', '__add__')  |
| -        | 11.5±0ms                   | 10.7±0ms                       |    0.94 | sparse.Arithmetic.time_arithmetic('coo', 'AA', 'multiply') |
| -        | 9.68±0ms                   | 9.13±0ms                       |    0.94 | sparse.Arithmetic.time_arithmetic('coo', 'AB', '__sub__')  |
| -        | 6.65±0ms                   | 6.27±0ms                       |    0.94 | sparse.Arithmetic.time_arithmetic('csr', 'AA', '__mul__')  |
| -        | 9.96±0ms                   | 9.30±0ms                       |    0.93 | sparse.Arithmetic.time_arithmetic('coo', 'BA', '__sub__')  |
| -        | 4.61±0ms                   | 4.28±0ms                       |    0.93 | sparse.Arithmetic.time_arithmetic('csr', 'AB', '__add__')  |
| -        | 10.4±0ms                   | 9.66±0ms                       |    0.93 | sparse.Arithmetic.time_arithmetic('dia', 'AB', 'multiply') |
| -        | 17.6±0ms                   | 16.1±0ms                       |    0.91 | sparse.Arithmetic.time_arithmetic('coo', 'AA', '__mul__')  |
| -        | 4.06±0ms                   | 3.64±0ms                       |    0.9  | sparse.Arithmetic.time_arithmetic('coo', 'BB', '__sub__')  |
| -        | 4.70±0ms                   | 4.20±0ms                       |    0.89 | sparse.Arithmetic.time_arithmetic('csr', 'AB', '__sub__')  |
| -        | 4.69±0ms                   | 4.20±0ms                       |    0.89 | sparse.Arithmetic.time_arithmetic('csr', 'BA', '__sub__')  |
| -        | 3.57±0ms                   | 3.16±0ms                       |    0.89 | sparse.Arithmetic.time_arithmetic('dia', 'BB', '__sub__')  |
| -        | 4.24±0ms                   | 3.69±0ms                       |    0.87 | sparse.Arithmetic.time_arithmetic('csc', 'BB', '__sub__')  |
| -        | 3.63±0ms                   | 3.14±0ms                       |    0.86 | sparse.Arithmetic.time_arithmetic('csr', 'BB', '__sub__')  |
| -        | 10.9±0ms                   | 9.19±0ms                       |    0.84 | sparse.Arithmetic.time_arithmetic('coo', 'AA', '__sub__')  |
| -        | 2.03±0ms                   | 1.54±0ms                       |    0.76 | sparse.Arithmetic.time_arithmetic('csr', 'AA', 'multiply') |
| -        | 998±0μs                    | 708±0μs                        |    0.71 | sparse.Arithmetic.time_arithmetic('csr', 'AA', '__sub__')  |