@perimosocordiae 2024-05-16
Here are my ASV results after removing the `--quick` flag and increasing the number of rounds to 4:

| Change   | Before [8ce393f1] <main>   | After [d9a04dc7] <fastbinop>   |   Ratio | Benchmark (Parameter)                                      |
|----------|----------------------------|--------------------------------|---------|------------------------------------------------------------|
| +        | 3.49±0.02ms                | 4.08±0.06ms                    |    1.17 | sparse.Arithmetic.time_arithmetic('csr', 'AB', 'multiply') |
| -        | 19.2±0.4ms                 | 18.2±0.4ms                     |    0.95 | sparse.Arithmetic.time_arithmetic('dia', 'AA', 'multiply') |
| -        | 14.5±0.5ms                 | 13.7±0.4ms                     |    0.94 | sparse.Arithmetic.time_arithmetic('coo', 'AA', '__mul__')  |
| -        | 6.21±0.03ms                | 5.81±0.03ms                    |    0.94 | sparse.Arithmetic.time_arithmetic('coo', 'AB', '__add__')  |
| -        | 1.34±0.02ms                | 1.26±0.01ms                    |    0.94 | sparse.Arithmetic.time_arithmetic('csc', 'AA', '__sub__')  |
| -        | 8.77±0.06ms                | 8.28±0.06ms                    |    0.94 | sparse.Arithmetic.time_arithmetic('dia', 'AB', '__sub__')  |
| -        | 1.00±0ms                   | 932±5μs                        |    0.93 | sparse.Arithmetic.time_arithmetic('csr', 'AA', '__add__')  |
| -        | 3.25±0.02ms                | 3.03±0.02ms                    |    0.93 | sparse.Arithmetic.time_arithmetic('csr', 'AB', '__add__')  |
| -        | 8.87±0.05ms                | 8.23±0.05ms                    |    0.93 | sparse.Arithmetic.time_arithmetic('dia', 'BA', '__sub__')  |
| -        | 6.24±0.04ms                | 5.76±0.02ms                    |    0.92 | sparse.Arithmetic.time_arithmetic('coo', 'AB', '__sub__')  |
| -        | 3.26±0.01ms                | 2.98±0.02ms                    |    0.91 | sparse.Arithmetic.time_arithmetic('csr', 'AB', '__sub__')  |
| -        | 7.27±0.3ms                 | 6.56±0.2ms                     |    0.9  | sparse.Arithmetic.time_arithmetic('coo', 'AA', '__sub__')  |
| -        | 6.37±0.03ms                | 5.73±0.04ms                    |    0.9  | sparse.Arithmetic.time_arithmetic('coo', 'BA', '__sub__')  |
| -        | 3.39±0.02ms                | 2.98±0.02ms                    |    0.88 | sparse.Arithmetic.time_arithmetic('csc', 'BA', '__sub__')  |
| -        | 3.41±0.02ms                | 2.99±0.01ms                    |    0.88 | sparse.Arithmetic.time_arithmetic('csr', 'BA', '__sub__')  |
| -        | 3.59±0.02ms                | 3.09±0.01ms                    |    0.86 | sparse.Arithmetic.time_arithmetic('coo', 'BB', '__sub__')  |
| -        | 3.60±0.05ms                | 3.09±0.01ms                    |    0.86 | sparse.Arithmetic.time_arithmetic('csc', 'BB', '__sub__')  |
| -        | 3.58±0.02ms                | 3.09±0.02ms                    |    0.86 | sparse.Arithmetic.time_arithmetic('csr', 'BB', '__sub__')  |
| -        | 3.58±0.02ms                | 3.08±0.01ms                    |    0.86 | sparse.Arithmetic.time_arithmetic('dia', 'BB', '__sub__')  |
| -        | 1.20±0.01ms                | 940±5μs                        |    0.79 | sparse.Arithmetic.time_arithmetic('csr', 'AA', 'multiply') |
| -        | 935±30μs                   | 614±4μs                        |    0.66 | sparse.Arithmetic.time_arithmetic('csr', 'AA', '__sub__')  |

Looks like the CSR-AB-multiply benchmark is the only real regression, so I'll take a look into why that's happening next.